requirements.txt
django-jalali==5.1.0
django==4.0.2
Pillow==9.0.0
djangorestframework==3.13.1